feat(entity_base): enhance set_psets method for decodable property sets (#10)
The set_psets method in the SupertypeIfcEntityOfPsetAndProp class has been updated. Now, it not only adds each property set name from the given dictionary to 'property_set_names' and assigns the corresponding value to an attribute with the same name but also checks if a property set name is decodable. If it is, the method decodes the name using the decode_2byte function and assigns the property set value to an attribute with the decoded name as well. This allows handling of property set names that require decoding.